+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* server/region_preview: Add (temporary!) endpoint for previewing regions
 * server: API should respond to POST requests as well as GET
 * server/region_preview: Read from stdin by default
 * docs: Make /docs deep-redirect to readthedocs, add link in help